% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_marginal_aic.R
\name{calc_marginal_aic}
\alias{calc_marginal_aic}
\title{Calculate marginal AIC}
\usage{
calc_marginal_aic(object)
}
\arguments{
\item{object}{An object of class \code{jointCPUE} returned by
\code{\link[=jointCPUE]{jointCPUE()}}.}
}
\value{
A one-row data.frame with columns \code{logLik_marginal}, \code{k},
and \code{AIC_marginal}.
}
\description{
Calculates an approximate marginal AIC by removing user-defined prior
penalties stored in \code{nll_prior} from the optimized objective value.
}
