% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R
\name{make_mesh}
\alias{make_mesh}
\title{Construct an SPDE mesh for jointCPUE}
\usage{
make_mesh(
  data_utm,
  xy_cols,
  type = c("kmeans", "cutoff", "tailored", "custom"),
  cutoff,
  n_knots,
  seed = 42,
  mesh = NULL,
  fmesher_func = fmesher::fm_rcdt_2d_inla,
  convex = NULL,
  concave = convex,
  ...
)
}
\arguments{
\item{data_utm}{A data.frame with utm coords.}

\item{xy_cols}{Character vector length 2 giving x/y column names in \code{data}.}

\item{type}{Mesh construction method.}

\item{cutoff}{Minimum allowed triangle edge length. Required for type="cutoff".}

\item{n_knots}{Number of knots if type="kmeans".}

\item{seed}{Seed for kmeans if type="kmeans".}

\item{mesh}{Optional custom mesh (required for type="custom").}

\item{fmesher_func}{Which fmesher function to use. Default is
\code{\link[fmesher:fm_rcdt_2d]{fmesher::fm_rcdt_2d_inla()}} (supports very simple cutoff meshes).}

\item{convex, concave}{If supplied, passed to \code{\link[fmesher:fm_nonconvex_hull]{fmesher::fm_nonconvex_hull()}}
and used as a boundary.}

\item{...}{Passed to \code{fmesher_func}. Examples: \code{max.edge}, \code{offset},
\code{cutoff}, \code{extend}, \code{refine}.}
}
\value{
An object of class \code{jointCPUEmesh} with elements \code{mesh}, \code{spde}, \code{A},
\code{loc_xy}, \code{xy_cols}, and \code{loc_centers} (for kmeans).
}
\description{
mesh constructor with convenience options:
\itemize{
\item "cutoff": simplest mesh given \code{cutoff}
\item "kmeans": place vertices using kmeans centers (\code{n_knots})
\item "tailored": build from all points but pass advanced args via \code{...}
\item "custom": supply your own \code{mesh}
}
}
