% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_q_diffs.R
\name{plot_q_diffs_system}
\alias{plot_q_diffs_system}
\title{Plot systematic q-differences}
\usage{
plot_q_diffs_system(object, fleet_labels = NULL)
}
\arguments{
\item{object}{A fitted \code{jointCPUE} object.}

\item{fleet_labels}{Optional character vector of length \code{n_f}.}
}
\value{
A ggplot object.
}
\description{
Plots fleet-specific systematic q-differences relative to the reference fleet.
}
