% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_parameters.R
\name{.make_parameters_jointCPUE}
\alias{.make_parameters_jointCPUE}
\title{Internal: Build initial parameter list for the jointCPUE TMB model}
\usage{
.make_parameters_jointCPUE(n_t, n_f, n_s)
}
\arguments{
\item{n_t}{Number of time steps.}

\item{n_f}{Number of fleets including the baseline fleet coded as 0.}

\item{n_s}{Number of SPDE mesh vertices.}
}
\value{
Named list of initial parameter values for TMB::MakeADFun().
}
\description{
Internal: Build initial parameter list for the jointCPUE TMB model
}
\keyword{internal}
