% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Prepare data objects and mesh for jointCPUE workflows}
\usage{
make_data(data_utm, mesh, area_scale = "auto")
}
\arguments{
\item{data_utm}{A data.frame containing required columns (with utm_x/y_scale)}

\item{mesh}{jointCPUEmesh built from make_mesh(), or a custom mesh}

\item{area_scale}{Numeric or "auto". Scaling factor for area_km2.}
}
\value{
A list with elements mesh, data, key, scales.
}
\description{
Main data-prep function for jointCPUE.
\itemize{
\item mesh/SPDE/A matrices
\item extrapolation key grid + areas
}
}
\author{
Rujia Bi \email{bikayla5@gmail.com}
}
